Use pre-generated RSA keys in tests
