+++ lib/auth/auth_test.go
@@ -184,6 +184,7 @@ func newAuthSuite(t *testing.T) *testPack {
 
 func TestMain(m *testing.M) {
 	utils.InitLoggerForTests()
+	native.PrecomputeTestKeys()
 	os.Exit(m.Run())
 }
